aichat/pyproject.toml
[project]
name = "aichat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.9.6",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "numpy>=2.4.2",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
]
